+++ src/Kernel/Console/Commands/MakeModel.php
@@ -2,10 +2,12 @@
 
 namespace Solital\Core\Kernel\Console\Commands;
 
+use Katrina\Katrina;
 use Solital\Core\Console\Command;
 use Solital\Core\Console\Interface\CommandInterface;
 use Solital\Core\Kernel\Application;
 use Solital\Core\Kernel\Console\HelpersTrait;
+use Nette\PhpGenerator\{ClassType, PhpNamespace, Property};
 
 class MakeModel extends Command implements CommandInterface
 {
@@ -35,13 +37,17 @@ class MakeModel extends Command implements CommandInterface
     public function handle(object $arguments, object $options): mixed
     {
         $model_dir = Application::getRootApp('Components/Model/', Application::DEBUG);
-        $model_template = Application::getConsoleComponent('ModelTemplate.php');
 
         if (isset($options->remove)) {
             $this->removeComponent($model_dir, $arguments->model_name . ".php");
         }
 
-        $res = $this->createComponent($model_template, $model_dir, $arguments->model_name);
+        $data = $this->codeGenerated($arguments->model_name);
+
+        $res = $this->createComponent($data, [
+            'component_name' => $arguments->model_name,
+            'directory' => $model_dir
+        ]);
 
         if ($res == true) {
             $this->success("Model successfully created!")->print()->break();
@@ -54,5 +60,43 @@ class MakeModel extends Command implements CommandInterface
         }
 
         return $this;
+    }
+
+    /**
+     * @param  mixed $model_name
+     * @return string
+     */
+    private function codeGenerated(string $model_name): string
+    {
+        $table_property = (new Property('table'))
+            ->setType('?string')
+            ->setValue("")
+            ->setProtected()
+            ->addComment("\n@var string|null\n");
+
+        $id_property = (new Property('id'))
+            ->setType('?string')
+            ->setValue("")
+            ->setProtected()
+            ->addComment("\n@var string|null\n");
+
+        $timestamp_property = (new Property('timestamp'))
+            ->setType('bool')
+            ->setValue(false)
+            ->setProtected()
+            ->addComment("\n@var bool\n");
+
+        $class = (new ClassType($model_name))
+            ->setExtends(Katrina::class)
+            ->addMember($table_property)
+            ->addMember($id_property)
+            ->addMember($timestamp_property)
+            ->addComment("@generated class generated using Vinci Console");
+
+        $data = (new PhpNamespace("Solital\Components\Model"))
+            ->add($class)
+            ->addUse(Katrina::class);
+
+        return $data;
     }
 }